image_distortion and line_detection
I destorted the picture and use canny-operator tu detect lines. It won't work in Openframeworks embedded OfxOpenCv but OpenCV itself. And the distorted picture seems weird because it's only distorted from one corner.
I'll try to take pictures with camera directly to see if it will be better.
But the canny-operator works very good in line-detection.